+++ nequip/dynamics/nequip_calculator.py
@@ -5,18 +5,13 @@ from ase.calculators.calculator import Calculator, all_changes
 
 from nequip.data import AtomicData, AtomicDataDict
 
-class NequIPCalculator(Calculator):
-    """ NequIP ASE Calculator. """
-    implemented_properties = ['energy', 'forces']
 
-    def __init__(
-            self,
-            predictor,
-            r_max,
-            device,
-            force_units_to_eva=1.,
-            **kwargs
-    ):
+class NequIPCalculator(Calculator):
+    """NequIP ASE Calculator."""
+
+    implemented_properties = ["energy", "forces"]
+
+    def __init__(self, predictor, r_max, device, force_units_to_eva=1.0, **kwargs):
         Calculator.__init__(self, **kwargs)
         self.results = {}
         self.predictor = predictor
@@ -24,12 +19,7 @@ class NequIPCalculator(Calculator):
         self.device = device
         self.force_units_to_eva = force_units_to_eva
 
-    def calculate(
-            self,
-            atoms=None,
-            properties=["energy"],
-            system_changes=all_changes
-    ):
+    def calculate(self, atoms=None, properties=["energy"], system_changes=all_changes):
         """
         Calculate properties.
 
@@ -42,20 +32,17 @@ class NequIPCalculator(Calculator):
         Calculator.calculate(self, atoms)
 
         # prepare data
-        data = AtomicData.from_ase(
-            atoms=atoms,
-            r_max=self.r_max
-        )
+        data = AtomicData.from_ase(atoms=atoms, r_max=self.r_max)
 
         data = data.to(self.device)
 
         # predict + extract data
         out = self.predictor(AtomicData.to_AtomicDataDict(data))
-        forces = out['forces'].detach().cpu().numpy()
-        energy = out['total_energy'].detach().cpu().item()
+        forces = out["forces"].detach().cpu().numpy()
+        energy = out["total_energy"].detach().cpu().item()
 
         # store results
         self.results = {
-            'energy': energy * self.force_units_to_eva,
-            'forces': forces * self.force_units_to_eva
+            "energy": energy * self.force_units_to_eva,
+            "forces": forces * self.force_units_to_eva,
         }