+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Most recent change on the bottom.
 
 ## [Unreleased]
+
+- Add MD w/ Nequip-ASE-calculator + run-MD script w/ custom Nose-Hoover
+
 ### Changed
 - Training now uses atomic file writes to avoid corruption if interupted